Fix deleting objects from S3
The UniteCMS\StorageBundle\Service\StorageService::deleteObject method
always used the us-east-1 region when initialising an S3Client object.
This meant that delete operations on buckets in any other region would
fail.

This commit changes the method to use $bucket_settings['region'] and
only fall back to us-east-1 if it's not set, allowing deletes to work.