Make S3 credentials optional
The `key` and `secret` bucket settings for file and image fields are
no longer required. If they're not specified, the S3 client will
automatically attempt to retrieve credentials from the EC2 instance
metadata server. This allows the EC2 instance role to be granted the
required S3 permissions, rather than having to add the key and secret
to the Unite domain configuration.